% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exports.R
\name{flot-exports}
\alias{\%>\%}
\alias{flot-exports}
\title{flot exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
to avoid listing the magrittr package
as depends of flot.
}

