% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flot.R
\name{flotChart}
\alias{flotChart}
\title{flot interactive plots}
\usage{
flotChart(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{data.table holding values to be plotted}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}
}
\value{
Interactive flot plot
}
\description{
R interface to interactive  plotting using the
\href{http://www.flotcharts.org/}{flot} JavaScript library.
}
\note{
See the \href{http://www.flotcharts.org/}{online documentation} for
additional details and examples.
}

