% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flotOptions.R
\name{flotOptions}
\alias{flotOptions}
\title{flot options}
\usage{
flotOptions(flotChart, ...)
}
\arguments{
\item{flotChart}{An object instantiated using \code{\link{flotChart}})
to add a series definition to.}

\item{legend}{list See \href{https://github.com/flot/flot/blob/master/API.md\#plot-options}{online documentation} (optional)}

\item{xaxis,yaxis,xaxes,yaxes}{list See \href{https://github.com/flot/flot/blob/master/API.md\#plot-options}{online documentation} (optional)}

\item{series}{list See \href{https://github.com/flot/flot/blob/master/API.md\#plot-options}{online documentation} (optional)}

\item{grid}{list See \href{https://github.com/flot/flot/blob/master/API.md\#plot-options}{online documentation} (optional)}

\item{tooltip}{logical See \href{https://github.com/krzysu/flot.tooltip}{online documentation} (optional)}

\item{tooltipOpts}{list See \href{https://github.com/krzysu/flot.tooltip}{online documentation} (optional)}

\item{selection}{list See \href{http://www.flotcharts.org/flot/examples/selection/}{online documentation} (optional)}
}
\value{
flotChart with additional series
}
\description{
Add options to a flot chart. All options are completely optional. They are documented individually
\href{https://github.com/flot/flot/blob/master/API.md\#plot-options}{online documentation} of the JS
library as accepted without validation and converted using RJSONIO::toJSON
}
\note{
See the
  \href{http://dcenergy.com}{online
  documentation} for additional details and examples.
}

