% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flotSeries.R
\name{flotSeries}
\alias{flotSeries}
\title{flot data series}
\usage{
flotSeries(flotChart, x = NULL, y = NULL, extra.cols = NULL,
  group = NULL, data = NULL, color = NULL, label = NULL, lines = NULL,
  bars = NULL, points = NULL, xaxis = NULL, yaxis = NULL,
  stack = NULL, clickable = F, hoverable = F, shadowSize = NULL,
  highlightColor = NULL)
}
\arguments{
\item{flotChart}{An object instantiated using \code{\link{flotChart}})
to add a series definition to.}

\item{x}{expression Evaluated in the context of the underlying data.table,
used to obtain values for the x-coordinate in the series}

\item{y}{expression Evaluated in the context of the underlying data.table,
used to obtain values for the y-coordinate in the series}

\item{extra.cols}{expression Possibly a vector of expressions evaluated in the context of the underlying data.table,
used to obtain values included in the data-series outside of the x- or
y-coordinate (optional)}

\item{group}{expression Evaluated in the context of hte underlying data.table, used to transform the data-set
from a long, to a wide data.table. (optional)}

\item{data}{data.table containing an alternative source for the data in this series. (optional)}

\item{color}{Color for series. The color is either a CSS color specification
(like "rgb(255, 100, 123)") or an integer that specifies which of auto-generated
colors to select, e.g. 0 will get color no. 0, etc. (optional)}

\item{label}{Label to display for series (uses names[2] if not supplied). (optional)}

\item{lines}{list See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{bars}{list See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{points}{list See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{xaxis}{numeric See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{yaxis}{numeric See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{clickable}{logical See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{hoverable}{logical See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{shadowSize}{numeric See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}

\item{highlightColor}{color or number See \href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} (optional)}
}
\value{
flotChart with additional series
}
\description{
Add a data series to a flot chart. Note that function call is necessary and sufficient
for at least one series to be charted on your plot.  This API allows you to customize
per-series parameters, overriding global settings specified using \code{\link{flotOptions}}).
All parameters of the single-series object as desribed in the
\href{https://github.com/flot/flot/blob/master/API.md\#data-format}{online documentation} of the JS
library as accepted without validation and converted using RJSONIO::toJSON
}
\note{
See the
  \href{http://dcenergy.com}{online
  documentation} for additional details and examples.
}

